# backend/swagger.yaml
openapi: 3.0.0
info:
  title: GigFlow API
  description: |
    Full-stack freelance marketplace API where Clients can post jobs (Gigs) and Freelancers can apply for them (Bids).
    
    ## Features
    - Secure authentication with JWT tokens in HttpOnly cookies
    - Gig management (CRUD operations)
    - Bidding system
    - Hiring logic with transactional integrity
    - Real-time notifications via Socket.io
    
    ## Authentication
    All protected endpoints require a valid JWT token stored in an HttpOnly cookie.
    The token is automatically set when you register or login.
  version: 1.0.0
  contact:
    name: GigFlow API Support
    email: support@gigflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/api
    description: Development server
  - url: https://api.gigflow.com/api
    description: Production server
...
paths:
...
  /bids:
    post:
      tags:
        - Bids
      summary: Submit a bid
      description: Creates a new bid on a gig (requires authentication, cannot bid on own gig)
      operationId: createBid
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - gigId
                - message
                - price
              properties:
                gigId:
                  type: string
                  example: 507f1f77bcf86cd799439011
                  description: ID of the gig to bid on
                message:
                  type: string
                  maxLength: 500
                  example: I have 5+ years of experience in web development and would love to work on this project.
                price:
                  type: number
                  minimum: 1
                  example: 4500
                  description: Bid amount in dollars
      responses:
        '201':
          description: Bid created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Bid'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bids/{gigId}:
    get:
      tags:
        - Bids
      summary: Get bids for a gig
      description: Retrieves all bids for a specific gig (only the gig owner can view)
      operationId: getBidsByGig
      security:
        - cookieAuth: []
      parameters:
        - name: gigId
          in: path
          required: true
          description: Gig ID
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Bids retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 5
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bid'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bids/{bidId}/hire:
    patch:
      tags:
        - Bids
      summary: Hire a freelancer
      description: |
        Hires a freelancer for a gig. This action:
        - Changes the selected bid status to "hired"
        - Changes the gig status from "open" to "assigned"
        - Automatically rejects all other pending bids for that gig
        - Sends real-time notifications via Socket.io
        
        **Important**: Uses MongoDB transactions to prevent race conditions.
        Only the gig owner can hire freelancers.
      operationId: hireFreelancer
      security:
        - cookieAuth: []
      parameters:
        - name: bidId
          in: path
          required: true
          description: Bid ID to hire
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Freelancer hired successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Freelancer hired successfully
                  data:
                    type: object
                    properties:
                      bid:
                        $ref: '#/components/schemas/Bid'
                      gig:
                        $ref: '#/components/schemas/Gig'
        '400':
          description: Bad request (gig already assigned, bid not available, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict - Gig has already been assigned to another freelancer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bids/user/my-bids:
    get:
      tags:
        - Bids
      summary: Get user's bids
      description: Retrieves all bids submitted by the authenticated user
      operationId: getUserBids
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User bids retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 3
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BidWithGig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HttpOnly cookie
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Manya Shukla
        email:
          type: string
          format: email
          example: shuklamanya99@gmail.com
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
...
    Gig:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        title:
          type: string
          example: Build a responsive e-commerce website
        description:
          type: string
          example: Looking for an experienced web developer...
        budget:
          type: number
          example: 5000
        ownerId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
          example: 507f1f77bcf86cd799439011
        status:
          type: string
          enum: [open, assigned]
          example: open
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
    Bid:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        gigId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Gig'
          example: 507f1f77bcf86cd799439011
        freelancerId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
          example: 507f1f77bcf86cd799439011
        message:
          type: string
          example: I have 5+ years of experience...
        price:
          type: number
          example: 4500
        status:
          type: string
          enum: [pending, hired, rejected]
          example: pending
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
    BidWithGig:
      allOf:
        - $ref: '#/components/schemas/Bid'
        - type: object
          properties:
            gigId:
              $ref: '#/components/schemas/Gig'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message describing what went wrong
  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Validation error message
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not authorized to access this route
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Internal Server Error